fix phpcs pointing
